backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
SQLAlchemy==2.1.0
psycopg2-binary==2.9.7
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic==1.10.12
python-dotenv==1.0.0
